working on a streaming file format for kd tree + lod
We first build a kd-tree as before, then chop it up into subtrees of
some desired depth limit to save out to disk. We could then look at
grouping these subtrees into files to aggregate levels of the tree together

+++ tools/rsf_file.cpp
@@ -58,12 +58,12 @@ void write_raw_surfels_v2(const std::string &fname, const std::vector<Surfel> &s
 	SplatKdTree kd_tree(bounds);
 
 	std::array<uint32_t, 4> header = {
-		packed_surfs.size(),
-		kd_tree.nodes.size() * sizeof(KdNode)
+		static_cast<uint32_t>(packed_surfs.size()),
+		static_cast<uint32_t>(kd_tree.nodes.size() * sizeof(KdNode)
 			+ (4 + kd_tree.primitive_indices.size()) * sizeof(uint32_t)
-			+ sizeof(Box),
-		kd_tree.nodes.size(),
-		kd_tree.primitive_indices.size()
+			+ sizeof(Box)),
+		static_cast<uint32_t>(kd_tree.nodes.size()),
+		static_cast<uint32_t>(kd_tree.primitive_indices.size())
 	};
 	fout.write(reinterpret_cast<const char*>(header.data()), sizeof(uint32_t) * header.size());
 	fout.write(reinterpret_cast<const char*>(&kd_tree.tree_bounds), sizeof(Box));
